Traditional First Python Program
It is traditional for your first program in a language to be one which writes "Hello World."  Simple as it may seem, it required 1) Downloading and installing the python interpreter, 2) Using a Python Integrated Development Environment that comes with Python, called IDLE, 3) Understanding the file structure in GitHub and creating a new repository.  Success :)

Although I am sure to be writing much more complicated code as time passes, I wanted to keep my original Hello World program for historical preservation.